FilteringAPI request changed and code moved to FoodResource class since it is more intuitive
